requirements.txt
rembg==2.0.72
pillow>=10.2.0
numpy>=1.26.0
